requirements.txt
ijson
pyyaml
requests
